requirements.txt
aiofiles==0.7.0
aioredis==2.0.0
aiosmtplib==1.1.6
aiosqlite==0.16.1
anyio==3.3.0
asgiref==3.4.1
async-timeout==3.0.1
atlastk==0.12.10
bcrypt==3.2.0
blinker==1.4
certifi==2021.5.30
cffi==1.14.6
click==8.0.1
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
fakeredis==1.5.2
fastapi==0.68.0
fastapi-mail==0.4.0
h11==0.12.0
httpcore==0.13.6
httptools==0.2.0
httpx==0.18.2
idna==3.2
importlib-metadata==4.6.3
iso8601==0.1.16
Jinja2==3.0.1
MarkupSafe==2.0.1
passlib==1.7.4
Pillow==8.3.1
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
pypika-tortoise==0.1.1
python-dotenv==0.19.0
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
starlette==0.14.2
tortoise==0.1.1
tortoise-orm==0.17.6
typing-extensions==3.10.0.0
uvicorn==0.14.0
watchgod==0.7
websockets==9.1
zipp==3.5.0
